+++ APICoches/src/main/java/com/example/demo/repository/DepartamentosRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Departamentos;
 
-public interface DepartamentosRepository extends JpaRepository<Departamentos, Integer>{
+public interface DepartamentosRepository extends JpaRepository<Departamentos, Long>{
 
 }
